+++ src/MediaPlayer/MediaPlayer.ts
@@ -1,7 +1,7 @@
 import { Player } from '..';
 import { Video } from '../types/Video';
 
-export type WrapperConstructor = new (player: Player) => Wrapper;
+export type MediaPlayerConstructor = new (player: Player) => MediaPlayer;
 
 export interface PlaybackSegment {
   /**
@@ -14,7 +14,7 @@ export interface PlaybackSegment {
   end?: number;
 }
 
-export interface Wrapper {
+export interface MediaPlayer {
   play: () => Promise<void>;
   pause: () => void;
   configureWithVideo: (video: Video, segment?: PlaybackSegment) => void;